autobulldoze: prepare for movement of auto-bulldoze code
When road/rail/wire tool is used, auto-bulldozing occurs before checking
the cost of the road/rail/wire being placed. This means when the user
is low on funds, the wire tool may end up bulldozing only, even though
there is not enough money for building what they want.

This patch prepares the way for moving where the auto-bulldoze check
is done in the case of the road/rail/wire tool.

git-svn-id: http://micropolis.googlecode.com/svn/trunk@588 d9718cc8-9f43-0410-858b-315f434eb58c